Quick test of elevation.
Made a new map and created an elevated piece of terrain. Placed a simple builder unit on a hill. Placed a ramp leading down. Clicked on the lower part of the map and the builder was able to navigate down the ramp. Currently there is an issue with placed buildings. They are clipping into the ground slightly causing any units to spawn to be also stuck in the ground.